feat: update css, date format, and create data components for get and spend point methods
