+++ src/pages/home/show.tsx
@@ -170,10 +170,10 @@ export const HomeShow = () => {
               {product.start_datetime && product.end_datetime
                 ? `${format(
                     new Date(product.start_datetime),
-                    "dd/MM/yyyy, hh:mm a"
+                    "dd MMM yyyy, HH:mm"
                   )} - ${format(
                     new Date(product.end_datetime),
-                    "dd/MM/yyyy, hh:mm a"
+                    "dd MMM yyyy, HH:mm"
                   )}`
                 : t("Date not available")}
             </p>
