+++ src/pages/login/index.tsx
@@ -5,7 +5,6 @@ import { loginWithLine } from "../../authProvider";
 import { supabase } from "../../lib/supabase";
 import { createTestSession } from "../../authProvider";
 import GlowfishIcon from "@/components/icons/GlowfishIcon";
-import { Link } from "react-router-dom";
 import LineIcon from "@/components/icons/LineIcon";
 import DrawerInfo from "@/components/company-info/DrawerInfo";
 
@@ -25,13 +24,11 @@ export const Login = () => {
   return (
     <section className="px-5 py-10 flex flex-col gap-20">
       <div className="flex flex-col gap-10">
-        {/* TODO: add GlowfishIcon */}
         <GlowfishIcon className="w-[102px] h-[58px]" />
         <h1 className="text-[31px] tracking-[0.43px] m-0">
           {t("Sign in to see all the event happening.")}
         </h1>
       </div>
-
       <div className="flex flex-col gap-5">
         <Button
           className="main-btn relative font-semibold"
@@ -40,7 +37,6 @@ export const Login = () => {
           <LineIcon className="w-7 h-7 absolute left-4" fill="#06c755" />
           {t("Continue with Line")}
         </Button>
-
         <DrawerInfo />
       </div>
     </section>
