+++ src/pages/products/index.tsx
@@ -132,7 +132,7 @@ export default function ProductsPage() {
           <Button
             variant="ghost"
             size="icon"
-            className="text-white h-12"
+            className="text-white h-12 !bg-transparent"
             onClick={() => navigate("/cart")}
           >
             <ShoppingCart className="h-6 w-6" />
